requirements.txt
streamlit
easyocr
pypdf2
pdf2image
pillow
requests
python-dotenv
chromadb
sentence-transformers
sqlite-utils
bcrypt==4.1.2